+++ docs/CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 ---------
 
+#### Next release
+- Changes:
+  - Remove coveralls plugin
+- Update dependencies:
+  - Kotlin to 1.5.30
+  - AndroidX Activity to 1.3.1
+  - Firebase Analytics to 19.0.1, Crashlytics to 18.2.1, Google Services plugin to 4.3.10
+  - Mockito to 3.12.3
+
 #### v0.17.1 (2021-08-02)
 - New features:
   - Add link to website